refactor: clean-up util functions. add logging
