+++ test/testUtils.ts
@@ -125,3 +125,21 @@ export async function cleanUpWorkspace(): Promise<any> {
 export function reloadConfiguration() {
   require('../src/configuration/configuration').configuration.reload();
 }
+
+/**
+ * Waits for the tabs to change after a command like 'gt' or 'gT' is run.
+ * Sometimes it is not immediate, so we must busy wait
+ * On certain versions, the tab changes are synchronous
+ * For those, a timeout is given
+ */
+export async function waitForTabChange(): Promise<void> {
+  await new Promise((resolve, reject) => {
+    setTimeout(resolve, 500);
+
+    const disposer = vscode.window.onDidChangeActiveTextEditor(textEditor => {
+      disposer.dispose();
+
+      resolve(textEditor);
+    });
+  });
+}
